Add stubs for python package as well as pipelines for testing (#35)
### Justification

This is the first step to creating a python package to contain details of examples that we don't want to include directly in notebooks. This change introduces the python package structure (directories), tests, and some poetry tasks that are used in GitHub actions for automatically testing on PR and submission

### Implementation

1. Add very basic python code as part of the initial package prototype. This includes the python code as well as the poetry setup for building (in `pyproject.toml`)
    2. Most of the poetry and pipeline information was reused from https://github.com/ni/nisystemlink-clients-python
3. Introduce `poe` tasks for 
    1. test using pytest
    2. linting using flake8 (also includes flake8 configuration)
    4. checking formatting using black
    5. automatically formatting using black
    6. checking types with mypy
7. Update CONTRIBUTING to talk about how
8. Added `python-package` workflow for GitHub to run tests on PR and submissions

### Testing

Added new python tests. These are mostly stubs but establish the tooling

## Checklist

- [x] I have read and followed the [Contributing Guidelines](../CONTRIBUTING.md)